+++ CHANGELOG.md
@@ -253,6 +253,7 @@
 - Tip de estado de Editar selección incluye Return.
 - Tip de estado de Renombrar selección incluye F2.
 - Tip de estado de Copiar ID de selección incluye Ctrl+Shift+C.
+- Tip de estado de Eliminar incluye Backspace o Delete.
 
 ### Cambiado